tests/wpt/web-platform-tests/app-history/navigate/state/updateCurrent/exception-order-not-fully-active-unserializablestate.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="iframe" src="/common/blank.html"></iframe>

<script>
async_test(t => {
  window.onload = t.step_func_done(() => {
    const wAppHistory = iframe.contentWindow.appHistory;
    const wDOMException = iframe.contentWindow.DOMException;

    iframe.remove();

    assert_throws_dom("InvalidStateError", wDOMException, () => {
      wAppHistory.updateCurrent({ state: document.body });
    });
    assert_equals(appHistory.current.getState(), undefined);
  });
}, `updateCurrent() with unserializable state while not fully active must throw an "InvalidStateError", not a "DataCloneError"`);
</script>
